Gamepad Metadata Renderer
Got the reactive part of the rendering working fairly nicely actually! I think the concept to look into now though is to split the updates into proper single state parts, because right now they all cross-update each other, which is a bit unfortunate.

I wonder if it makes sense to abstract the underlying Gamepad API into an event-based one, and then add SolidJS on top of that, for rendering only, or to just use Solid's state tools to manage it on it's own? I thought that second one could do that very nicely, but it still seems to have the issues I kind of thought it would have. Dang every time I work with this API, it's just not quite what I'd expect it to work like. That's okay though! It takes time :) There's a reason they built it this way I think.

I also learned that my SteelSeries Nimbus shows as two separate controllers because one is for the home button only, and the rest are the regular controller buttons. Definitely interesting.