GameController.js Demo
This is still kind of a big mix of gunk, it's not working how I'd like it to.

To get it working, you also have to cd into the node_modules folder for the new controller lib, and build it with `npm run build`. This commit is more for a conceptual setup of things.